Partialview-structure-validation
Controllerlar daha uygun bir şekilde ayrıldı.
Viewların bazıları partial view olacak şekilde ayarlandı.
Ufak olaylar(hava durumu gibi) Partial action(?) olacak şekilde adapte
edildi.
Javascript validation'ları iyileştirildi.
