+++ gemi/Controllers/GemiController.cs
@@ -254,42 +254,6 @@ namespace gemi.Controllers
             return View(Records);
         }
 
-        [HttpPost]
-        public ActionResult GetShipsByName(string ship_name)
-        {
-            ShipData shipData = new ShipData();
-            TanimData tanimData = new TanimData();
-            Dictionary<int,string> tanimlar = tanimData.GetTanimlar(); //id,string dict
-            int ship_id = tanimlar.FirstOrDefault(x => x.Value == ship_name).Key;
-            ViewBag.tanimlar = tanimlar;
-            List<Ship> ships = shipData.GetShipsByName(ship_id);
-            return View(ships);
-        }
-
-        [HttpPost]
-        public ActionResult GetShipsByDate(string begin, string end)
-        {
-            ShipData shipData = new ShipData();
-            TanimData tanimData = new TanimData();
-            Dictionary<int, string> tanimlar = tanimData.GetTanimlar(); //id,string dict
-            ViewBag.tanimlar = tanimlar;
-            List<Ship> ships = shipData.GetShipsBetweenDate(Convert.ToDateTime(begin + ",00:00:00").Date, Convert.ToDateTime(end + ",00:00:00").Date);
-
-            return View(ships);
-        }
-
-        [HttpPost]
-        public ActionResult GetShipsByNameAndDate(string ship_name, string begin, string end)
-        {
-            ShipData shipData = new ShipData();
-            TanimData tanimData = new TanimData();
-            Dictionary<int, string> tanimlar = tanimData.GetTanimlar();
-            int ship_id = tanimlar.FirstOrDefault(x => x.Value == ship_name).Key;
-            ViewBag.tanimlar = tanimlar;
-
-            List<Ship> ships = shipData.GetShipsByDateAndName(ship_id, Convert.ToDateTime(begin + ",00:00:00").Date, Convert.ToDateTime(end + ",00:00:00").Date);
-
-            return View(ships);
-        }
+        
     }
 }
