+++ gemi/Views/Admin/UserPanel.cshtml
@@ -87,7 +87,6 @@
     <input type="submit" />
 </form></div>
 
-
 <div class="Users">
 @foreach(gemi.Entities.Roles role in (List<gemi.Entities.Roles>)ViewBag.roles)
 {
