--- gemi/Views/Home/Search.cshtml
@@ -1,45 +1,0 @@
-﻿@{
-    ViewBag.Title = "Search";
-}
-
-<script>
-    function ValidateSearchString(object) {
-        var result;
-        var frm = $(object).parent().parent();
-        if (frm.find("input:eq(0)").val() == "") {
-            CannotBeEmpty(frm.find("input:eq(0)"));
-            result = false;
-        }
-        else if (frm.find("input:eq(0)").val() == "") {
-            TooLong(frm.find("input:eq(0)"), 10);
-            result = false;
-        }
-        return result;
-    }
-</script>
-<fieldset>
-<legend>Referans ile arama</legend>
-<form id="search-form" action="/Gemi/GetShip" method="get">
-    <div style="width: 70%">
-        <input type="text" name="ref_id" class="required-search" value="Referans numarasını girin" onclick="ClickToEmpty(this)"/>
-    </div>
-    <div style="width: 25%">
-        <input type="submit" onclick="return ValidateSearch(this)" />
-    </div>
-</form>
-</fieldset>
-<fieldset>
-<legend>Gemi adı ile arama</legend>
-<form action="/Gemi/GetShipsByName" method="post">
-    <input type="text" name="ship_name" value="Gemi adı giriniz" onclick="ClickToEmpty(this)"/>
-    <input type="submit" onclick="return ValidateSearchString(this)"/>
-</form>
-</fieldset>
-<fieldset>
-<legend>Tarih aralığı ile arama</legend>
-<form action="/Gemi/GetShipsByDate" method="post">
-    <input class="datepick" type="text" name="begin" readonly value="Başlangıç tarihi"/>
-    <input class="datepick" type="text" name="end" readonly value="Bitiş tarihi"/>
-    <input type="submit" />
-</form>
-</fieldset>
