+++ gemi/Views/Search/Index.cshtml
@@ -1,0 +1,7 @@
+﻿@{
+    ViewBag.Title = "Search";
+}
+
+<a href="~/Search/GetShipsByRef">Referans ile ara</a><br />
+<a href="~/Search/GetShipsByName">İsim ile ara</a><br />
+<a href="~/Search/GetShipsByDate">Tarih aralığı ile ara</a><br />
