+++ gemi/Views/Search/SearchRef.cshtml
@@ -1,0 +1,17 @@
+﻿@{
+    ViewBag.Title = "GetShipsByRef";
+}
+
+<h2>GetShipsByRef</h2>
+
+<fieldset>
+<legend>Referans ile arama</legend>
+<form id="search-form" action="/Gemi/GetShip" method="get">
+    <div style="width: 70%">
+        <input type="text" name="ref_id" class="required-search" value="Referans numarasını girin" onclick="ClickToEmpty(this)"/>
+    </div>
+    <div style="width: 25%">
+        <input type="submit" onclick="return ValidateSearch(this)" />
+    </div>
+</form>
+</fieldset>
