+++ gemi/Views/Search/_ShipInfo.cshtml
@@ -1,0 +1,1 @@
+﻿<div> @ViewBag.tanimlar[Model.shipId] </div> @Model.time.ToString().Remove(10) <div><a href="~/Gemi/GetShip?ref_id=@Model.refId">@Model.refId</a></div><br />
