+++ gemi/Views/Shared/_Currency.cshtml
@@ -1,10 +1,4 @@
-﻿@{
-    ViewBag.Title = "ShowCurrency";
-}
-
-<h2>ShowCurrency</h2>
-
-@model Dictionary<string,string>
+﻿@model Dictionary<string,string>
 <table>
 @foreach (KeyValuePair<string, string> i in Model)
 {
