+++ gemi/Views/Shared/_Forecast.cshtml
@@ -1,0 +1,6 @@
+﻿<table>
+    <tr><td>Şehir:</td><td>@Model["Sehir"]</td></tr>
+    <tr><td>Hava:</td><td>@Model["Hava"]</td></tr>
+    <tr><td>Sıcaklık:</td><td>@Model["Sıcaklık"]</td></tr>
+    <tr><td>Güncel:</td><td>@Model["Zaman"]</td></tr>
+</table>
