+++ gemi/Web.config
@@ -19,7 +19,7 @@
     <add key="UnobtrusiveJavaScriptEnabled" value="true" />
   </appSettings>
   <system.web>
-    <httpRuntime targetFramework="4.5.1" />
+    <httpRuntime targetFramework="4.5.1" maxRequestLength="1048576"/>
     <compilation debug="true" targetFramework="4.5.1" />
     <authentication mode="Forms">
       <forms loginUrl="~/Account/Login" timeout="2880" />
